+++ bmcd/src/app.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 pub mod bmc_application;
 pub mod bmc_info;
+pub mod config_export;
 pub mod cooling_device;
 pub mod event_application;
 pub mod firmware_catalog;